health-bot/actions/requirements.txt
rasa-sdk==3.6.2
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
twilio>=9.0.0
uvloop>=0.19.0; sys_platform != 'win32'

